added nav and resolved style error
